+++ services/service.go
@@ -1,13 +1,8 @@
 package services
 
 import (
-	"crypto/sha256"
-	"crypto/sha512"
 	"encoding/binary"
-	"fmt"
-	"hash"
 	"hash/crc32"
-	"math"
 	"strconv"
 	"url-shortener/configs"
 	"url-shortener/models"
@@ -22,13 +17,13 @@ func IsExistByBoomFilter(config configs.GlobalConfig, longUrl string) bool {
 func GenShortUrlHash(config configs.GlobalConfig, longUrl string) (string, error) {
 	// CRC32 here
 	data := []byte(longUrl)
-	hash := crc32.NewIEEE()
+	hashFunc := crc32.NewIEEE()
 
-	_, err := hash.Write(data)
+	_, err := hashFunc.Write(data)
 	if err != nil {
 		return "", err
 	}
-	checksum := hash.Sum32()
+	checksum := hashFunc.Sum32()
 
 	// Convert uint32 to byte slice
 	byteSlice := make([]byte, 4)
@@ -47,25 +42,4 @@ func IsMappedLongToShortDB(config configs.GlobalConfig, longUrl string) bool {
 	//TODO: implement this function
 	// Check in DB
 	return false
-}
-
-func BoomFilterApply() {
-	bloomFilter = initBloomFilter()
-	var s1 = "hoale"
-	bloomFilter.Set(s1)
-	fmt.Println(bloomFilter.Check(s1))
-	fmt.Println(bloomFilter.Check("hoalethe"))
-
-}
-
-func initBloomFilter() models.BloomFilters {
-	hashFunctions := []hash.Hash{
-		sha256.New224(),
-		sha256.New(),
-		sha512.New(),
-		sha512.New384(),
-	}
-
-	bloomFilter = models.CreateBloomFilter(hashFunctions, math.MaxUint32)
-	return bloomFilter
 }
